refactor: 내부 열거형을 외부 클래스로 이동
