+++ tomcat/src/main/java/com/java/http/request_response/HttpResponse.java
@@ -34,22 +34,6 @@ public record HttpResponse(
         );
     }
 
-    public enum StatusCode {
-        OK("OK", 200),
-        FOUND("Found", 302),
-        NOT_FOUND("Not Found", 404),
-        INTERNAL_SERVER_ERROR("Internal Server Error", 500),
-        ;
-
-        final String codeName;
-        final int codeNumber;
-
-        StatusCode(String codeName, int codeNumber) {
-            this.codeName = codeName;
-            this.codeNumber = codeNumber;
-        }
-    }
-
     public byte[] toByteArray() {
         try (ByteArrayOutputStream baos = new ByteArrayOutputStream()) {
             baos.write("%s %s %s\r\n".formatted(version, statusCode.codeNumber, statusCode.codeName).getBytes(StandardCharsets.UTF_8));
