+++ tomcat/src/main/java/com/techcourse/servlet_impl/LoginPageServlet.java
@@ -6,8 +6,8 @@ import com.java.http.request_response.HttpResponse;
 import com.java.http.session.Session;
 import com.java.servlet.Servlet;
 
-import static com.java.http.request_response.HttpRequest.HttpMethod.GET;
-import static com.java.http.request_response.HttpResponse.StatusCode.OK;
+import static com.java.http.request_response.HttpMethod.GET;
+import static com.java.http.request_response.StatusCode.OK;
 
 public class LoginPageServlet implements Servlet {
