+++ tomcat/src/main/java/com/techcourse/servlet_impl/LoginServlet.java
@@ -13,7 +13,7 @@ import org.slf4j.LoggerFactory;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.java.http.request_response.HttpRequest.HttpMethod.POST;
+import static com.java.http.request_response.HttpMethod.POST;
 
 public class LoginServlet implements Servlet {
